Fix integration test compilation issues
Agent-Logs-Url: https://github.com/TopperDEL/uplink.NET.New/sessions/0759f2f5-d95e-4254-9805-8d5ebde11406

+++ tests/uplink.NET.IntegrationTests/MultipartUploadServiceTests.cs
@@ -217,7 +217,7 @@ public class MultipartUploadServiceTests
                 partBytes);
 
             Assert.True(string.IsNullOrEmpty(partResult.Error), partResult.Error);
-            Assert.Equal(length, partResult.BytesWritten);
+            Assert.Equal((uint)length, partResult.BytesWritten);
             uploadedPartCount++;
         }
